Shorted GeneratePatient() method + Deleted PatientGenerator
